requirements.txt
pytest==9.0.2
playwright==1.58.0
pytest-playwright==0.7.2
requests==2.31.0
pytest-html==4.2.0
allure-pytest==2.15.3
python-dotenv==1.0.1